extra_requirements.txt
jsonlines==1.2.0
pandas==1.0.3
seaborn==0.10.1
ipython==7.14.0
pytorch-transformers==1.2.0
torch>=0.4.0
sklearn
spacy-pytorch-transformers[cuda100]
